+++ zerohertzLib/vision/data.py
@@ -22,8 +22,6 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-import math
-import multiprocessing as mp
 import os
 import shutil
 import urllib.parse
@@ -32,13 +30,11 @@ from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 
 import cv2
 import numpy as np
-from numpy.typing import DTypeLike, NDArray
 from tqdm import tqdm
 
-from zerohertzLib.util import Json, JsonDir, rmtree, write_json
+from zerohertzLib.util import Json, rmtree, write_json
 
-from .convert import poly2cwh, poly2mask, poly2xyxy, xyxy2poly
-from .visual import bbox, mask
+from .convert import poly2cwh, poly2xyxy, xyxy2poly
 
 
 def _get_image_paths(path: str) -> List[str]:
@@ -58,342 +54,6 @@ def _get_image_paths(path: str) -> List[str]:
     for ext_ in ext:
         image_paths += glob(os.path.join(path, f"*.{ext_}"))
     return image_paths
-
-
-class ImageLoader:
-    """경로와 image의 수를 지정하여 경로 내 image를 return하는 class
-
-    Args:
-        path (``Optional[str]``): Image들이 존재하는 경로
-        cnt (``Optional[int]``): 호출 시 return 할 image의 수
-
-    Attributes:
-        image_paths (``List[str]``): 지정한 경로 내 image들의 경로
-
-    Methods:
-        __len__:
-            Returns:
-                ``int``: ``cnt`` 에 해당하는 image들의 수
-
-        __getitem__:
-            Args:
-                idx (``int``): 입력 index
-
-            Returns:
-                ``Union[Tuple[str, NDArray[np.uint8]], Tuple[List[str], List[NDArray[np.uint8]]]``: ``cnt`` 에 따른 file 경로 및 image 값
-
-    Examples:
-        >>> il = zz.vision.ImageLoader()
-        >>> len(il)
-        510
-        >>> il[0][0]
-        './1.2.410.200001.1.9999.1.20220513101953581.1.1.jpg'
-        >>> il[0][1].shape
-        (480, 640, 3)
-        >>> il = zz.vision.ImageLoader(cnt=4)
-        >>> len(il)
-        128
-        >>> il[0][0]
-        ['./1.2.410.200001.1.9999.1.20220513101953581.1.1.jpg', '...', '...', '...']
-        >>> il[0][1][0].shape
-        (480, 640, 3)
-        >>> len(il[0][0])
-        4
-        >>> len(il[0][1])
-        4
-    """
-
-    def __init__(self, path: Optional[str] = "./", cnt: Optional[int] = 1) -> None:
-        self.cnt = cnt
-        self.image_paths = _get_image_paths(path)
-        self.image_paths.sort()
-
-    def __len__(self) -> int:
-        return math.ceil(len(self.image_paths) / self.cnt)
-
-    def __getitem__(
-        self, idx: int
-    ) -> Union[
-        Tuple[str, NDArray[np.uint8]], Tuple[List[str], List[NDArray[np.uint8]]]
-    ]:
-        if self.cnt == 1:
-            return (
-                self.image_paths[idx],
-                cv2.imread(self.image_paths[idx], cv2.IMREAD_UNCHANGED),
-            )
-        return (
-            self.image_paths[self.cnt * idx : self.cnt * (idx + 1)],
-            [
-                cv2.imread(path, cv2.IMREAD_UNCHANGED)
-                for path in self.image_paths[self.cnt * idx : self.cnt * (idx + 1)]
-            ],
-        )
-
-
-class JsonImageLoader:
-    """JSON file을 통해 image와 JSON file 내 정보를 불러오는 class
-
-    Args:
-        data_path (``str``): 목표 data가 존재하는 directory 경로
-        json_path (``str``): 목표 JSON file이 존재하는 directory 경로
-        json_key (``str``): ``data_path`` 에서 data의 file 이름을 나타내는 key 값
-
-    Attributes:
-        json (``zerohertzLib.util.JsonDir``): JSON file들을 읽어 data 구축 시 활용
-
-    Methods:
-        __len__:
-            Returns:
-                ``int``: 읽어온 JSON file들의 수
-
-        __getitem__:
-            읽어온 JSON file들을 list와 같이 indexing 후 해당하는 image return
-
-            Args:
-                idx (``int``): 입력 index
-
-            Returns:
-                ``Tuple[NDArray[np.uint8], zerohertzLib.util.Json]``: Image와 JSON 내 정보
-
-    Examples:
-        >>> jil = zz.vision.JsonImageLoader(data_path, json_path, json_key)
-        100%|█████████████| 17248/17248 [00:04<00:00, 3581.22it/s]
-        >>> img, js = jil[10]
-        >>> img.shape
-        (600, 800, 3)
-        >>> js.tree()
-        └─ info
-            └─ name
-            └─ date_created
-        ...
-    """
-
-    def __init__(
-        self,
-        data_path: str,
-        json_path: str,
-        json_key: str,
-    ) -> None:
-        self.data_path = data_path
-        self.json_path = json_path
-        self.json = JsonDir(json_path)
-        self.json_key = self.json._get_key(json_key)
-
-    def __len__(self) -> int:
-        return len(self.json)
-
-    def __getitem__(self, idx: int) -> Tuple[NDArray[np.uint8], Json]:
-        data_name = self.json[idx].get(self.json_key)
-        img = cv2.imread(os.path.join(self.data_path, data_name), cv2.IMREAD_UNCHANGED)
-        return img, self.json[idx]
-
-
-class YoloLoader:
-    """YOLO format의 dataset을 읽고 시각화하는 class
-
-    Args:
-        data_path (``Optional[str]``): Image가 존재하는 directory 경로
-        txt_path (``Optional[str]``): YOLO format의 ``.txt`` 가 존재하는 directory 경로
-        poly (``Optional[bool]``): ``.txt`` file의 format (``False``: detection, ``True``: segmentation)
-        absolute (``Optional[bool]``): ``.txt`` file의 절대 좌표계 여부 (``False``: relative coordinates, ``True``: absolute coordinates)
-        vis_path (``Optional[str]``): 시각화 image들이 저장될 경로
-        class_color (``Optional[Dict[Union[int, str], Tuple[int]]]``): 시각화 결과에 적용될 class에 따른 색상
-
-    Methods:
-        __len__:
-            Returns:
-                ``int``: 읽어온 image file들의 수
-
-        __getitem__:
-            Index에 따른 image와 ``.txt`` file에 대한 정보 return (``vis_path`` 와 ``class_color`` 입력 시 시각화 image ``vis_path`` 에 저장)
-
-            Args:
-                idx (``int``): 입력 index
-
-            Returns:
-                ``Tuple[NDArray[np.uint8], List[int], List[NDArray[DTypeLike]]]``: 읽어온 image와 그에 따른 ``class_list`` 및 ``bbox`` 혹은 ``poly``
-
-    Examples:
-        >>> data_path = ".../images"
-        >>> txt_path = ".../labels"
-        >>> class_color = {0: (0, 255, 0), 1: (255, 0, 0), 2: (0, 0, 255)}
-        >>> yolo = YoloLoader(data_path, txt_path, poly=True, absolute=False, vis_path="tmp", class_color=class_color)
-        >>> image, class_list, objects = yolo[0]
-        >>> type(image)
-        <class 'numpy.ndarray'>
-        >>> class_list
-        [1, 1]
-        >>> len(objects)
-        2
-    """
-
-    def __init__(
-        self,
-        data_path: Optional[str] = "images",
-        txt_path: Optional[str] = "labels",
-        poly: Optional[bool] = True,
-        absolute: Optional[bool] = False,
-        vis_path: Optional[str] = None,
-        class_color: Optional[Dict[Union[int, str], Tuple[int]]] = None,
-    ) -> None:
-        self.data_path = data_path
-        self.data_paths = _get_image_paths(self.data_path)
-        self.txt_path = txt_path
-        self.poly = poly
-        self.absolute = absolute
-        self.vis_path = vis_path
-        if vis_path is not None:
-            if class_color is None:
-                raise ValueError(
-                    "Visualization requires the 'class_color' variable to be specified"
-                )
-            rmtree(vis_path)
-            self.class_color = class_color
-
-    def __len__(self) -> int:
-        return len(self.data_paths)
-
-    def __getitem__(
-        self, idx: int
-    ) -> Tuple[NDArray[np.uint8], List[int], List[NDArray[DTypeLike]]]:
-        data_path = self.data_paths[idx]
-        data_file_name = data_path.split("/")[-1]
-        txt_path = os.path.join(
-            self.txt_path, ".".join(data_file_name.split(".")[:-1]) + ".txt"
-        )
-        img = cv2.imread(data_path)
-        try:
-            class_list, objects = self._convert(txt_path, img)
-        except FileNotFoundError:
-            print(f"'{data_file_name}' is not found")
-            return None, None, None
-        if self.vis_path is not None:
-            self._visualization(data_file_name, img, class_list, objects)
-        return img, class_list, objects
-
-    def _convert(
-        self, txt_path: str, img: NDArray[np.uint8]
-    ) -> Tuple[List[int], List[NDArray[DTypeLike]]]:
-        class_list = []
-        objects = []
-        with open(txt_path, "r", encoding="utf-8") as file:
-            data_lines = file.readlines()
-        for data_line in data_lines:
-            data_str = data_line.strip().split(" ")
-            class_list.append(int(data_str[0]))
-            if self.poly:
-                obj = np.array(list(map(float, data_str[1:]))).reshape(-1, 2)
-                if not self.absolute:
-                    obj *= img.shape[:2][::-1]
-            else:
-                obj = np.array(list(map(float, data_str[1:])))
-                if not self.absolute:
-                    obj *= img.shape[:2][::-1] * 2
-            objects.append(obj)
-        return class_list, objects
-
-    def _visualization(
-        self,
-        file_name: str,
-        img: NDArray[np.uint8],
-        class_list: List[int],
-        objects: List[NDArray[DTypeLike]],
-    ) -> None:
-        mks = np.zeros((len(objects), *img.shape[:2]), bool)
-        if self.poly:
-            for idx, poly in enumerate(objects):
-                mks[idx] = poly2mask(poly, img.shape[:2])
-            img = mask(img, mks, class_list=class_list, class_color=self.class_color)
-        else:
-            for idx, (cls, box) in enumerate(zip(class_list, objects)):
-                img = bbox(img, box, self.class_color[cls])
-        cv2.imwrite(os.path.join(self.vis_path, file_name), img)
-
-    def _value(self, img, obj, labels, cls):
-        original_height, original_width = img.shape[:2]
-        obj *= 100
-        if self.poly:
-            obj /= (original_width, original_height)
-            return {
-                "original_width": original_width,
-                "original_height": original_height,
-                "image_rotation": 0,
-                "value": {
-                    "points": obj.tolist(),
-                    "closed": True,
-                    "polygonlabels": [labels[cls]],
-                },
-                "from_name": "label",
-                "to_name": "image",
-                "type": "polygonlabels",
-                "origin": "manual",
-            }
-        obj[:2] -= obj[2:] / 2
-        obj /= (original_width, original_height) * 2
-        return {
-            "original_width": original_width,
-            "original_height": original_height,
-            "image_rotation": 0,
-            "value": {
-                "x": obj[0],
-                "y": obj[1],
-                "width": obj[2],
-                "height": obj[3],
-                "rectanglelabels": [labels[cls]],
-            },
-            "from_name": "label",
-            "to_name": "image",
-            "type": "rectanglelabels",
-            "origin": "manual",
-        }
-
-    def _annotation(self, args) -> Dict[str, Any]:
-        idx, directory, labels = args
-        img, class_list, objects = self[idx]
-        data_path = self.data_paths[idx]
-        data_file_name = data_path.split("/")[-1]
-        annotation = {
-            "data": {"image": f"data/local-files/?d={directory}/{data_file_name}"}
-        }
-        result_data = []
-        for cls, obj in zip(class_list, objects):
-            result_data.append(self._value(img, obj, labels, cls))
-        annotation["annotations"] = [{"result": result_data}]
-        return annotation
-
-    def labelstudio(
-        self,
-        directory: Optional[str] = "image",
-        labels: Optional[List[str]] = None,
-        mp_num: Optional[int] = 0,
-    ) -> None:
-        """
-        YOLO format의 data를 Label Studio에서 확인 및 수정할 수 있게 변환
-
-        Args:
-            directory (``Optional[str]``): Label Studio 내 ``/home/user/{directory}`` 의 이름
-            labels (``Optional[List[str]]``): YOLO format의 ``.txt`` 상에서 index에 따른 label의 이름
-            mp_num (``Optional[int]``): 병렬 처리에 사용될 process의 수 (``0``: 직렬 처리)
-
-        Returns:
-            ``None``: ``{path}.json`` 으로 결과 저장
-
-        Examples
-            >>> yolo.labelstudio("images", mp_num=10, labels=["t1", "t2", "t3", "t4"])
-        """
-        if labels is None:
-            labels = [str(i) for i in range(100)]
-        json_data = []
-        if mp_num == 0:
-            for idx in range(len(self)):
-                json_data.append(self._annotation([idx, directory, labels]))
-        else:
-            args = [[idx, directory, labels] for idx in range(len(self))]
-            with mp.Pool(processes=mp_num) as pool:
-                annotations = pool.map(self._annotation, args)
-            for annotation in annotations:
-                json_data.append(annotation)
-        write_json(json_data, self.data_path)
 
 
 class LabelStudio:
@@ -665,10 +325,10 @@ class LabelStudio:
         """
         if label is None:
             label = []
-        os.makedirs(os.path.join(target_path, "images"), exist_ok=True)
-        os.makedirs(os.path.join(target_path, "labels"), exist_ok=True)
+        rmtree(os.path.join(target_path, "images"))
+        rmtree(os.path.join(target_path, "labels"))
         for file_path, result in tqdm(self):
-            img_file_name = file_path.split("/")[-1]
+            img_file_name = os.path.basename(file_path)
             txt_file_name = ".".join(img_file_name.split(".")[:-1]) + ".txt"
             converted_gt = []
             for lab, poly in zip(result["labels"], result["polys"]):
@@ -716,8 +376,8 @@ class LabelStudio:
         """
         if label is None:
             label = {}
-        os.makedirs(os.path.join(target_path, "images"), exist_ok=True)
-        os.makedirs(os.path.join(target_path, "labels"), exist_ok=True)
+        rmtree(os.path.join(target_path, "images"))
+        rmtree(os.path.join(target_path, "labels"))
         for file_path, result in tqdm(self):
             img_file_name = file_path.split("/")[-1]
             json_file_name = ".".join(img_file_name.split(".")[:-1])
